+++ kombu/transport/qpid.py
@@ -5,16 +5,19 @@ kombu.transport.qpid
 `Qpid`_ transport using `qpid-python`_ as the client and `qpid-tools`_ for
 broker management.
 
-.. _`Qpid`: http://qpid.apache.org/
-.. _`qpid-python`: http://pypi.python.org/pypi/qpid-python/
-.. _`qpid-tools`: http://pypi.python.org/pypi/qpid-tools/
-.. _`Issue 2199`: https://github.com/celery/celery/issues/2199
-
 The use this transport you must install the necessary dependencies. These
 dependencies are available via PyPI and can be installed using the pip
 command:
 
-`pip install qpid-tools qpid-python`
+.. code-block:: console
+
+    $ pip install kombu[qpid]
+
+or to install the requirements manually:
+
+.. code-block:: console
+
+    $ pip install qpid-tools qpid-python
 
 .. admonition:: Python 3 and PyPy Limitations
 
@@ -22,10 +25,9 @@ command:
     to underlying dependencies not being compatible. This version is
     tested and works with with Python 2.7.
 
-.. admonition:: Potential Deadlock
-
-    This transport should be used with caution due to a known
-    potential deadlock. See `Issue 2199`_ for more details.
+.. _`Qpid`: http://qpid.apache.org/
+.. _`qpid-python`: http://pypi.python.org/pypi/qpid-python/
+.. _`qpid-tools`: http://pypi.python.org/pypi/qpid-tools/
 
 Authentication
 ==============
@@ -73,7 +75,7 @@ The :attr:`~kombu.Connection.transport_options` argument to the
 :class:`~kombu.Connection` object are passed directly to the
 :class:`qpid.messaging.endpoints.Connection` as keyword arguments. These
 options override and replace any other default or specified values. If using
-Celery with Kombu, this can be accomplished by setting the
+Celery, this can be accomplished by setting the
 *BROKER_TRANSPORT_OPTIONS* Celery option.
 
 """
@@ -84,10 +86,9 @@ import select
 import socket
 import ssl
 import sys
-import threading
 import time
+import uuid
 
-from itertools import count
 from gettext import gettext as _
 
 import amqp.protocol
@@ -154,19 +155,24 @@ class AuthenticationFailure(Exception):
 class QoS(object):
     """A helper object for message prefetch and ACKing purposes.
 
+    :keyword prefetch_count: Initial prefetch count, hard set to 1.
+    :type prefetch_count: int
+
+
     NOTE: prefetch_count is currently hard set to 1, and needs to be improved
 
-    This object is instantiated 1-for-1 with a :class:`Channel`. QoS
-    allows prefetch_count to be set to the number of outstanding messages
-    the corresponding :class:`Channel` should be allowed to prefetch.
-    Setting prefetch_count to 0 disables prefetch limits, and the object
-    can hold an arbitrary number of messages.
+    This object is instantiated 1-for-1 with a
+    :class:`~.kombu.transport.qpid.Channel` instance. QoS allows
+    ``prefetch_count`` to be set to the number of outstanding messages
+    the corresponding :class:`~kombu.transport.qpid.Channel` should be
+    allowed to prefetch.  Setting ``prefetch_count`` to 0 disables
+    prefetch limits, and the object can hold an arbitrary number of messages.
 
     Messages are added using :meth:`append`, which are held until they are
     ACKed asynchronously through a call to :meth:`ack`. Messages that are
     received, but not ACKed will not be delivered by the broker to another
     consumer until an ACK is received, or the session is closed. Messages
-    are referred to using delivery_tag integers, which are unique per
+    are referred to using delivery_tag, which are unique per
     :class:`Channel`. Delivery tags are managed outside of this object and
     are passed in with a message to :meth:`append`. Un-ACKed messages can
     be looked up from QoS using :meth:`get` and can be rejected and
@@ -175,19 +181,13 @@ class QoS(object):
     """
 
     def __init__(self, session, prefetch_count=1):
-        """Instantiate a QoS object.
-
-        :keyword prefetch_count: Initial prefetch count, hard set to 1.
-        :type prefetch_count: int
-
-        """
         self.session = session
         self.prefetch_count = 1
         self._not_yet_acked = OrderedDict()
 
     def can_consume(self):
-        """Return True if the :class:`Channel` can consume more messages,
-        else False.
+        """Return True if the :class:`~kombu.transport.qpid.Channel` can
+        consume more messages, else False.
 
         Used to ensure the client adheres to currently active prefetch
         limits.
@@ -196,53 +196,58 @@ class QoS(object):
             without violating the prefetch_count. If prefetch_count is 0,
             can_consume will always return True.
         :rtype: bool
+
         """
-        return not self.prefetch_count or len(self._not_yet_acked) < self\
-            .prefetch_count
+        return (
+            not self.prefetch_count or
+            len(self._not_yet_acked) < self.prefetch_count
+        )
 
     def can_consume_max_estimate(self):
         """Return the remaining message capacity for the associated
-        :class:`Channel`.
+        :class:`kombu.transport.qpid.Channel`.
 
         Returns an estimated number of outstanding messages that a
-        :class:`Channel` can accept without exceeding prefetch_count. If
-        prefetch_count is 0, then this method returns 1.
+        :class:`kombu.transport.qpid.Channel` can accept without
+        exceeding ``prefetch_count``. If ``prefetch_count`` is 0, then
+        this method returns 1.
 
         :returns: The number of estimated messages that can be fetched
             without violating the prefetch_count.
         :rtype: int
+
         """
-        if self.prefetch_count:
-            return self.prefetch_count - len(self._not_yet_acked)
-        else:
-            return 1
+        return 1 if not self.prefetch_count else (
+            self.prefetch_count - len(self._not_yet_acked)
+        )
 
     def append(self, message, delivery_tag):
         """Append message to the list of un-ACKed messages.
 
-        Add a message, referenced by the integer delivery_tag, for ACKing,
+        Add a message, referenced by the delivery_tag, for ACKing,
         rejecting, or getting later. Messages are saved into an
-        :class:`~kombu.utils.compat.OrderedDict` by delivery_tag.
+        :class:`collections.OrderedDict` by delivery_tag.
 
         :param message: A received message that has not yet been ACKed.
         :type message: qpid.messaging.Message
-        :param delivery_tag: An integer number to refer to this message by
+        :param delivery_tag: A UUID to refer to this message by
             upon receipt.
-        :type delivery_tag: int
+        :type delivery_tag: uuid.UUID
+
         """
         self._not_yet_acked[delivery_tag] = message
 
     def get(self, delivery_tag):
-        """
-        Get an un-ACKed message by delivery_tag. If called with an invalid
-        delivery_tag a KeyError is raised.
+        """Get an un-ACKed message by delivery_tag. If called with an invalid
+        delivery_tag a :exc:`KeyError` is raised.
 
         :param delivery_tag: The delivery tag associated with the message
             to be returned.
-        :type delivery_tag: int
+        :type delivery_tag: uuid.UUID
 
         :return: An un-ACKed message that is looked up by delivery_tag.
         :rtype: qpid.messaging.Message
+
         """
         return self._not_yet_acked[delivery_tag]
 
@@ -254,7 +259,8 @@ class QoS(object):
 
         :param delivery_tag: the delivery tag associated with the message
             to be acknowledged.
-        :type delivery_tag: int
+        :type delivery_tag: uuid.UUID
+
         """
         message = self._not_yet_acked.pop(delivery_tag)
         self.session.acknowledge(message=message)
@@ -262,7 +268,7 @@ class QoS(object):
     def reject(self, delivery_tag, requeue=False):
         """Reject a message by delivery_tag.
 
-        Explicitly notify the broker that the :class:`Channel` associated
+        Explicitly notify the broker that the channel associated
         with this QoS object is rejecting the message that was previously
         delivered.
 
@@ -272,12 +278,13 @@ class QoS(object):
 
         :param delivery_tag: The delivery tag associated with the message
             to be rejected.
-        :type delivery_tag: int
+        :type delivery_tag: uuid.UUID
         :keyword requeue: If True, the broker will be notified to requeue
             the message. If False, the broker will be told to drop the
             message entirely. In both cases, the message will be removed
             from this object.
         :type requeue: bool
+
         """
         message = self._not_yet_acked.pop(delivery_tag)
         QpidDisposition = qpid.messaging.Disposition
@@ -291,7 +298,13 @@ class QoS(object):
 class Channel(base.StdChannel):
     """Supports broker configuration and messaging send and receive.
 
-    A Channel object is designed to have method-parity with a Channel as
+    :param connection: A Connection object that this Channel can
+        reference. Currently only used to access callbacks.
+    :type connection: kombu.transport.qpid.Connection
+    :param transport: The Transport this Channel is associated with.
+    :type transport: kombu.transport.qpid.Transport
+
+    A channel object is designed to have method-parity with a Channel as
     defined in AMQP 0-10 and earlier, which allows for the following broker
     actions:
 
@@ -307,26 +320,26 @@ class Channel(base.StdChannel):
     Channels are designed to all share a single TCP connection with a
     broker, but provide a level of isolated communication with the broker
     while benefiting from a shared TCP connection. The Channel is given
-    its :class:`Connection` object by the :class:`Transport` that
-    instantiates the Channel.
+    its :class:`~kombu.transport.qpid.Connection` object by the
+    :class:`~kombu.transport.qpid.Transport` that
+    instantiates the channel.
 
-    This Channel inherits from :class:`~kombu.transport.base.StdChannel`,
-    which makes this a 'native' Channel versus a 'virtual' Channel which
+    This channel inherits from :class:`~kombu.transport.base.StdChannel`,
+    which makes this a 'native' channel versus a 'virtual' channel which
     would inherit from :class:`kombu.transports.virtual`.
 
-    Messages sent using this Channel are assigned a delivery_tag. The
+    Messages sent using this channel are assigned a delivery_tag. The
     delivery_tag is generated for a message as they are prepared for
     sending by :meth:`basic_publish`. The delivery_tag is unique per
-    Channel instance using :meth:`~itertools.count`. The delivery_tag has
-    no meaningful context in other objects, and is only maintained in the
-    memory of this object, and the underlying :class:`QoS` object that
-    provides support.
+    channel instance. The delivery_tag has no meaningful context in other
+    objects, and is only maintained in the memory of this object, and the
+    underlying :class:`QoS` object that provides support.
 
-    Each Channel object instantiates exactly one :class:`QoS` object for
+    Each channel object instantiates exactly one :class:`QoS` object for
     prefetch limiting, and asynchronous ACKing. The :class:`QoS` object is
     lazily instantiated through a property method :meth:`qos`. The
     :class:`QoS` object is a supporting object that should not be accessed
-    directly except by the Channel itself.
+    directly except by the channel itself.
 
     Synchronous reads on a queue are done using a call to :meth:`basic_get`
     which uses :meth:`_get` to perform the reading. These methods read
@@ -375,18 +388,7 @@ class Channel(base.StdChannel):
     #: Binary <-> ASCII codecs.
     codecs = {'base64': Base64()}
 
-    #: counter used to generate delivery tags for this channel.
-    _delivery_tags = count(1)
-
     def __init__(self, connection, transport):
-        """Instantiate a Channel object.
-
-        :param connection: A Connection object that this Channel can
-            reference. Currently only used to access callbacks.
-        :type connection: Connection
-        :param transport: The Transport this Channel is associated with.
-        :type transport: Transport
-        """
         self.connection = connection
         self.transport = transport
         qpid_connection = connection.get_qpid_connection()
@@ -403,9 +405,9 @@ class Channel(base.StdChannel):
         from a queue by name. This method creates a
         :class:`~qpid.messaging.endpoints.Receiver` to read from the queue
         using the :class:`~qpid.messaging.endpoints.Session` saved on the
-        associated :class:`Transport`. The receiver is closed before the
-        method exits. If a message is available, a
-        :class:`qpid.messaging.Message` object is returned. If no message is
+        associated :class:`~kombu.transport.qpid.Transport`.  The receiver
+        is closed before the method exits. If a message is available, a
+        :class:`qpid.messaging.Message` object is returned.  If no message is
         available, a :class:`qpid.messaging.exceptions.Empty` exception is
         raised.
 
@@ -419,6 +421,7 @@ class Channel(base.StdChannel):
         :rtype: :class:`qpid.messaging.Message`
         :raises: :class:`qpid.messaging.exceptions.Empty` if no
                  message is available.
+
         """
         rx = self.transport.session.receiver(queue)
         try:
@@ -443,8 +446,8 @@ class Channel(base.StdChannel):
         This method creates a :class:`qpid.messaging.endpoints.Sender` to
         send the message to the queue using the
         :class:`qpid.messaging.endpoints.Session` created and referenced by
-        the associated :class:`Transport`. The sender is closed before the
-        method exits.
+        the associated :class:`~kombu.transport.qpid.Transport`.  The sender
+        is closed before the method exits.
 
         External calls for put functionality should be done using
         :meth:`basic_publish`.
@@ -460,10 +463,11 @@ class Channel(base.StdChannel):
             should be sent on. If no exchange is specified, the message is
             sent directly to a queue specified by routing_key.
         :type exchange: str
+
         """
         if not exchange:
-            address = '%s; {assert: always, node: {type: queue}}' % \
-                      routing_key
+            address = '%s; {assert: always, node: {type: queue}}' % (
+                routing_key,)
             msg_subject = None
         else:
             address = '%s/%s; {assert: always, node: {type: topic}}' % (
@@ -509,6 +513,7 @@ class Channel(base.StdChannel):
 
         :raises: :class:`qpid.messaging.exceptions.NotFound` if the queue
                  being purged cannot be found.
+
         """
         queue_to_purge = self._broker.getQueue(queue)
         if queue_to_purge is None:
@@ -532,6 +537,7 @@ class Channel(base.StdChannel):
 
         :return the number of messages in the queue specified by name.
         :rtype: int
+
         """
         queue_to_check = self._broker.getQueue(queue)
         message_depth = queue_to_check.values['msgDepth']
@@ -550,6 +556,7 @@ class Channel(base.StdChannel):
 
         :param queue: The name of the queue to be deleted.
         :type queue: str
+
         """
         self._purge(queue)
         self._broker.delQueue(queue)
@@ -563,6 +570,7 @@ class Channel(base.StdChannel):
         :return: True if a queue exists on the broker, and false
             otherwise.
         :rtype: bool
+
         """
         if self._broker.getQueue(queue):
             return True
@@ -678,6 +686,7 @@ class Channel(base.StdChannel):
         :keyword if_empty: If True, only delete the queue if it is empty. If
             False, delete the queue if it is empty or not.
         :type if_empty: bool
+
         """
         if self._has_queue(queue):
             if if_empty and self._size(queue):
@@ -713,6 +722,7 @@ class Channel(base.StdChannel):
         :keyword durable: True if the exchange should be durable, or False
             otherwise.
         :type durable: bool
+
         """
         options = {'durable': durable}
         try:
@@ -726,6 +736,7 @@ class Channel(base.StdChannel):
 
         :param exchange_name: The name of the exchange to be deleted.
         :type exchange_name: str
+
         """
         self._broker.delExchange(exchange_name)
 
@@ -745,6 +756,7 @@ class Channel(base.StdChannel):
         :param routing_key: The bind key that the specified queue should
             bind to the specified exchange with.
         :type routing_key: str
+
         """
         self._broker.bind(exchange, queue, routing_key)
 
@@ -766,6 +778,7 @@ class Channel(base.StdChannel):
         :param routing_key: The existing bind key between the specified
             queue and a specified exchange that should be unbound.
         :type routing_key: str
+
         """
         self._broker.unbind(exchange, queue, routing_key)
 
@@ -798,6 +811,7 @@ class Channel(base.StdChannel):
 
         :raises: :class:`qpid.messaging.exceptions.NotFound` if the queue
                  being purged cannot be found.
+
         """
         return self._purge(queue)
 
@@ -828,6 +842,7 @@ class Channel(base.StdChannel):
 
         :return: The received message.
         :rtype: :class:`~kombu.transport.virtual.Message`
+
         """
         try:
             qpid_message = self._get(queue)
@@ -851,7 +866,8 @@ class Channel(base.StdChannel):
 
         :param delivery_tag: The delivery tag associated with the message
             to be acknowledged.
-        :type delivery_tag: int
+        :type delivery_tag: uuid.UUID
+
         """
         self.qos.ack(delivery_tag)
 
@@ -869,7 +885,7 @@ class Channel(base.StdChannel):
 
         :param delivery_tag: The delivery tag associated with the message
             to be rejected.
-        :type delivery_tag: int
+        :type delivery_tag: uuid.UUID
         :keyword requeue: If False, the rejected message will be dropped by
             the broker and not delivered to any other consumers. If True,
             then the rejected message will be requeued for delivery to
@@ -910,10 +926,9 @@ class Channel(base.StdChannel):
         handled by the caller of :meth:`~Transport.drain_events`. Messages
         can be ACKed after being received through a call to :meth:`basic_ack`.
 
-        If no_ack is True, the messages are immediately ACKed to avoid a
-        memory leak in qpid.messaging when messages go un-ACKed. The no_ack
-        flag indicates that the receiver of the message does not intent to
-        call :meth:`basic_ack`.
+        If no_ack is True, The no_ack flag indicates that the receiver of
+        the message will not call :meth:`basic_ack` later. Since the
+        message will not be ACKed later, it is ACKed immediately.
 
         :meth:`basic_consume` transforms the message object type prior to
         calling the callback. Initially the message comes in as a
@@ -940,6 +955,7 @@ class Channel(base.StdChannel):
         :param consumer_tag: a tag to reference the created consumer by.
             This consumer_tag is needed to cancel the consumer.
         :type consumer_tag: an immutable object
+
         """
         self._tag_to_queue[consumer_tag] = queue
 
@@ -949,9 +965,7 @@ class Channel(base.StdChannel):
             delivery_tag = message.delivery_tag
             self.qos.append(qpid_message, delivery_tag)
             if no_ack:
-                # Celery will not ack this message later, so we should to
-                # avoid a memory leak in qpid.messaging due to un-ACKed
-                # messages.
+                # Celery will not ack this message later, so we should ack now
                 self.basic_ack(delivery_tag)
             return callback(message)
 
@@ -973,6 +987,7 @@ class Channel(base.StdChannel):
             cancelled. Originally specified when the consumer was created
             as a parameter to :meth:`basic_consume`.
         :type consumer_tag: an immutable object
+
         """
         if consumer_tag in self._receivers:
             receiver = self._receivers.pop(consumer_tag)
@@ -987,6 +1002,7 @@ class Channel(base.StdChannel):
         known consumer_tag. It also closes the self._broker sessions. Closing
         the sessions implicitly causes all outstanding, un-ACKed messages to
         be considered undelivered by the broker.
+
         """
         if not self.closed:
             self.closed = True
@@ -1005,6 +1021,7 @@ class Channel(base.StdChannel):
 
         :return: An already existing, or newly created QoS object
         :rtype: :class:`QoS`
+
         """
         if self._qos is None:
             self._qos = self.QoS(self.transport.session)
@@ -1021,6 +1038,7 @@ class Channel(base.StdChannel):
 
         :param prefetch_count: Not used. This method is hard-coded to 1.
         :type prefetch_count: int
+
         """
         self.qos.prefetch_count = 1
 
@@ -1055,6 +1073,7 @@ class Channel(base.StdChannel):
             attributes. See parameters for more details on attributes that
             can be set.
         :rtype: dict
+
         """
         properties = properties or {}
         info = properties.setdefault('delivery_info', {})
@@ -1077,7 +1096,7 @@ class Channel(base.StdChannel):
         - wraps the body as a buffer object, so that
             :class:`qpid.messaging.endpoints.Sender` uses a content type
             that can support arbitrarily large messages.
-        - assigns a delivery_tag generated through self._delivery_tags
+        - sets delivery_tag to a random uuid.UUID
         - sets the exchange and routing_key info as delivery_info
 
         Internally uses :meth:`_put` to send the message synchronously. This
@@ -1095,6 +1114,7 @@ class Channel(base.StdChannel):
         :param routing_key: The routing key to be used as the message is
             submitted onto the exchange.
         :type routing_key: str
+
         """
         message['body'], body_encoding = self.encode_body(
             message['body'], self.body_encoding,
@@ -1103,7 +1123,7 @@ class Channel(base.StdChannel):
         props = message['properties']
         props.update(
             body_encoding=body_encoding,
-            delivery_tag=next(self._delivery_tags),
+            delivery_tag=uuid.uuid4(),
         )
         props['delivery_info'].update(
             exchange=exchange,
@@ -1131,6 +1151,7 @@ class Channel(base.StdChannel):
             encoding used. If encoding is not specified, the body is passed
             through unchanged.
         :rtype: tuple
+
         """
         if encoding:
             return self.codecs.get(encoding).encode(body), encoding
@@ -1154,6 +1175,7 @@ class Channel(base.StdChannel):
         :return: If encoding is specified, the decoded body is returned.
             If encoding is not specified, the body is returned unchanged.
         :rtype: str
+
         """
         if encoding:
             return self.codecs.get(encoding).decode(body)
@@ -1176,6 +1198,7 @@ class Channel(base.StdChannel):
 
         :return: The exchange type either 'direct', 'topic', or 'fanout'.
         :rtype: str
+
         """
         qpid_exchange = self._broker.getExchange(exchange)
         if qpid_exchange:
@@ -1186,12 +1209,45 @@ class Channel(base.StdChannel):
 
 
 class Connection(object):
-    """Encapsulate a connection object for the :class:`Transport`.
+    """Encapsulate a connection object for the
+    :class:`~kombu.transport.qpid.Transport`.
 
-    A Connection object is created by a :class:`Transport` during a call to
-    :meth:`Transport.establish_connection`. The :class:`Transport` passes in
+    :param host: The host that connections should connect to.
+    :param port: The port that connection should connect to.
+    :param username: The username that connections should connect with.
+        Optional.
+    :param password: The password that connections should connect with.
+        Optional but requires a username.
+    :param transport: The transport type that connections should use.
+        Either 'tcp', or 'ssl' are expected as values.
+    :param timeout: the timeout used when a Connection connects
+        to the broker.
+    :param sasl_mechanisms: The sasl authentication mechanism type to use.
+        refer to SASL documentation for an explanation of valid
+        values.
+
+    .. note::
+
+        qpid.messaging has an AuthenticationFailure exception type, but
+        instead raises a ConnectionError with a message that indicates an
+        authentication failure occurred in those situations.
+        ConnectionError is listed as a recoverable error type, so kombu
+        will attempt to retry if a ConnectionError is raised. Retrying
+        the operation without adjusting the credentials is not correct,
+        so this method specifically checks for a ConnectionError that
+        indicates an Authentication Failure occurred. In those
+        situations, the error type is mutated while preserving the
+        original message and raised so kombu will allow the exception to
+        not be considered recoverable.
+
+
+    A connection object is created by a
+    :class:`~kombu.transport.qpid.Transport` during a call to
+    :meth:`~kombu.transport.qpid.Transport.establish_connection`.  The
+    :class:`~kombu.transport.qpid.Transport` passes in
     connection options as keywords that should be used for any connections
-    created. Each :class:`Transport` creates exactly one Connection.
+    created. Each :class:`~kombu.transport.qpid.Transport` creates exactly
+    one Connection.
 
     A Connection object maintains a reference to a
     :class:`~qpid.messaging.endpoints.Connection` which can be accessed
@@ -1214,46 +1270,13 @@ class Connection(object):
     All keyword arguments are collected into the connection_options dict
     and passed directly through to
     :meth:`qpid.messaging.endpoints.Connection.establish`.
+
     """
 
     # A class reference to the :class:`Channel` object
     Channel = Channel
 
     def __init__(self, **connection_options):
-        """Instantiate a Connection object.
-
-        The following parameters are expected:
-
-        * host: The host that connections should connect to.
-        * port: The port that connection should connect to.
-        * username: The username that connections should connect with.
-              Optional.
-        * password: The password that connections should connect with.
-              Optional but requires a username.
-        * transport: The transport type that connections should use.
-              Either 'tcp', or 'ssl' are expected as values.
-        * timeout: the timeout used when a Connection connects to the
-              broker.
-        * sasl_mechanisms: The sasl authentication mechanism type to use.
-              refer to SASL documentation for an explanation of valid
-              values.
-
-        Creates a :class:`qpid.messaging.endpoints.Connection` object with
-        the saved parameters, and stores it as _qpid_conn.
-
-        qpid.messaging has an AuthenticationFailure exception type, but
-        instead raises a ConnectionError with a message that indicates an
-        authentication failure occurred in those situations.
-        ConnectionError is listed as a recoverable error type, so kombu
-        will attempt to retry if a ConnectionError is raised. Retrying
-        the operation without adjusting the credentials is not correct,
-        so this method specifically checks for a ConnectionError that
-        indicates an Authentication Failure occurred. In those
-        situations, the error type is mutated while preserving the
-        original message and raised so kombu will allow the exception to
-        not be considered recoverable.
-
-        """
         self.connection_options = connection_options
         self.channels = []
         self._callbacks = {}
@@ -1303,6 +1326,7 @@ class Connection(object):
 
         :return: The existing qpid.messaging.Connection
         :rtype: :class:`qpid.messaging.endpoints.Connection`
+
         """
         return self._qpid_conn
 
@@ -1311,17 +1335,19 @@ class Connection(object):
 
         Closing the connection will close all associated session, senders, or
         receivers used by the Connection.
+
         """
         self._qpid_conn.close()
 
     def close_channel(self, channel):
         """Close a Channel.
 
-        Close a channel specified by a reference to the :class:`Channel`
-        object.
+        Close a channel specified by a reference to the
+        :class:`~kombu.transport.qpid.Channel` object.
 
         :param channel: Channel that should be closed.
-        :type channel: Channel
+        :type channel: :class:`~kombu.transport.qpid.Channel`.
+
         """
         try:
             self.channels.remove(channel)
@@ -1329,81 +1355,6 @@ class Connection(object):
             pass
         finally:
             channel.connection = None
-
-
-class ReceiversMonitor(threading.Thread):
-    """A monitoring thread that reads and handles messages from all receivers.
-
-    A single instance of ReceiversMonitor is expected to be created by
-    :class:`Transport`.
-
-    In :meth:`monitor_receivers`, the thread monitors all receivers
-    associated with the session created by the Transport using the blocking
-    call to session.next_receiver(). When any receiver has messages
-    available, a symbol '0' is written to the self._w_fd file descriptor. The
-    :meth:`monitor_receivers` is designed not to exit, and loops over
-    session.next_receiver() forever.
-
-    The entry point of the thread is :meth:`run` which calls
-    :meth:`monitor_receivers`.
-
-    The thread is designed to be daemonized, and will be forcefully killed
-    when all non-daemon threads have already exited.
-    """
-
-    def __init__(self, session, w):
-        """Instantiate a ReceiversMonitor object
-
-        :param session: The session which needs all of its receivers
-            monitored.
-        :type session: :class:`qpid.messaging.endpoints.Session`
-        :param w: The file descriptor to write the '0' into when
-            next_receiver unblocks.
-        :type w: int
-        """
-        super(ReceiversMonitor, self).__init__()
-        self._session = session
-        self._w_fd = w
-
-    def run(self):
-        """Thread entry point for ReceiversMonitor
-
-        Calls :meth:`monitor_receivers` with a log-and-reenter behavior for
-        non connection errors. This guards against unexpected exceptions
-        which could cause this thread to exit unexpectedly.
-
-        A :class:`qpid.messaging.exceptions.SessionClosed` exception should
-        cause this thread to exit. This is a normal exit condition and the
-        thread is no longer needed.
-
-        If a connection error occurs, the exception needs to be propagated
-        to MainThread where the kombu exception handler can properly handle
-        it. The exception is stored as saved_exception on the self._session
-        object. The character 'e' is then written to the self.w_fd file
-        descriptor and then this thread exits.
-        """
-        while True:
-            try:
-                self.monitor_receivers()
-            except Transport.recoverable_connection_errors as exc:
-                self._session.saved_exception = exc
-                os.write(self._w_fd, 'e')
-                break
-            except SessionClosed:
-                break
-            except Exception as exc:
-                logger.error(exc, exc_info=1)
-            time.sleep(10)
-
-    def monitor_receivers(self):
-        """Monitor all receivers, and write to _w_fd when a message is ready.
-
-        The call to next_receiver() blocks until a message is ready. Once a
-        message is ready, write a '0' to _w_fd.
-        """
-        while True:
-            self._session.next_receiver()
-            os.write(self._w_fd, '0')
 
 
 class Transport(base.Transport):
@@ -1439,6 +1390,7 @@ class Transport(base.Transport):
     For backwards compatibility to the pre Kombu 3.0 exception interface, the
     recoverable errors are also listed as `connection_errors` and
     `channel_errors`.
+
     """
 
     # Reference to the class that should be used as the Connection object
@@ -1474,12 +1426,6 @@ class Transport(base.Transport):
     channel_errors = recoverable_channel_errors
 
     def __init__(self, *args, **kwargs):
-        """Instantiate a Transport object.
-
-        This method creates a pipe, and saves the read and write file
-        descriptors as attributes. The behavior of the read file descriptor
-        is modified to be non-blocking using fcntl.fcntl.
-        """
         self.verify_runtime_environment()
         super(Transport, self).__init__(*args, **kwargs)
         self.r, self._w = os.pipe()
@@ -1499,6 +1445,7 @@ class Transport(base.Transport):
         RuntimeError is raised.
 
         :raises: RuntimeError if the runtime environment is not acceptable.
+
         """
         if getattr(sys, 'pypy_version_info', None):
             raise RuntimeError(
@@ -1523,28 +1470,30 @@ class Transport(base.Transport):
                 'with your package manager. You can also try `pip install '
                 'qpid-python`.')
 
+    def _qpid_message_ready_handler(self, session):
+        os.write(self._w, '0')
+
+    def _qpid_async_exception_notify_handler(self, obj_with_exception, exc):
+        os.write(self._w, 'e')
+
     def on_readable(self, connection, loop):
         """Handle any messages associated with this Transport.
 
         This method clears a single message from the externally monitored
         file descriptor by issuing a read call to the self.r file descriptor
         which removes a single '0' character that was placed into the pipe
-        by :class:`ReceiversMonitor`. Once a '0' is read, all available
-        events are drained through a call to :meth:`drain_events`.
+        by the Qpid session message callback handler. Once a '0' is read,
+        all available events are drained through a call to
+        :meth:`drain_events`.
 
         The behavior of self.r is adjusted in __init__ to be non-blocking,
         ensuring that an accidental call to this method when no more messages
         will arrive will not cause indefinite blocking.
 
-        If the self.r file descriptor receives the character 'e', an error
-        occurred in the background thread, and this thread should raise the
-        saved exception. The exception is stored as saved_exception on the
-        session object.
-
         Nothing is expected to be returned from :meth:`drain_events` because
         :meth:`drain_events` handles messages by calling callbacks that are
-        maintained on the :class:`Connection` object. When
-        :meth:`drain_events` returns, all associated messages have been
+        maintained on the :class:`~kombu.transport.qpid.Connection` object.
+        When :meth:`drain_events` returns, all associated messages have been
         handled.
 
         This method calls drain_events() which reads as many messages as are
@@ -1570,14 +1519,13 @@ class Transport(base.Transport):
         :param connection: The connection associated with the readable
             events, which contains the callbacks that need to be called for
             the readable objects.
-        :type connection: Connection
+        :type connection: kombu.transport.qpid.Connection
         :param loop: The asynchronous loop object that contains epoll like
             functionality.
         :type loop: kombu.async.Hub
+
         """
-        symbol = os.read(self.r, 1)
-        if symbol == 'e':
-            raise self.session.saved_exception
+        os.read(self.r, 1)
         try:
             self.drain_events(connection)
         except socket.timeout:
@@ -1589,7 +1537,7 @@ class Transport(base.Transport):
         Register the callback self.on_readable to be called when an
         external epoll loop sees that the file descriptor registered is
         ready for reading. The file descriptor is created by this Transport,
-        and is updated by the ReceiversMonitor thread.
+        and is written to when a message is available.
 
         Because supports_ev == True, Celery expects to call this method to
         give the Transport an opportunity to register a read file descriptor
@@ -1604,9 +1552,10 @@ class Transport(base.Transport):
 
         :param connection: A reference to the connection associated with
             this Transport.
-        :type connection: Connection
+        :type connection: kombu.transport.qpid.Connection
         :param loop: A reference to the external loop.
         :type loop: kombu.async.hub.Hub
+
         """
         loop.add_reader(self.r, self.on_readable, connection, loop)
 
@@ -1630,6 +1579,7 @@ class Transport(base.Transport):
 
         :return: The created :class:`Connection` object is returned.
         :rtype: :class:`Connection`
+
         """
         conninfo = self.client
         for name, default_value in items(self.default_connection_params):
@@ -1685,17 +1635,23 @@ class Transport(base.Transport):
         conn = self.Connection(**opts)
         conn.client = self.client
         self.session = conn.get_qpid_connection().session()
-        monitor_thread = ReceiversMonitor(self.session, self._w)
-        monitor_thread.daemon = True
-        monitor_thread.start()
+        self.session.set_message_received_notify_handler(
+            self._qpid_message_ready_handler
+        )
+        conn.get_qpid_connection().set_async_exception_notify_handler(
+            self._qpid_async_exception_notify_handler
+        )
+        self.session.set_async_exception_notify_handler(
+            self._qpid_async_exception_notify_handler
+        )
         return conn
 
     def close_connection(self, connection):
-        """
-        Close the :class:`Connection` object.
+        """Close the :class:`Connection` object.
 
         :param connection: The Connection that should be closed.
         :type connection: :class:`kombu.transport.qpid.Connection`
+
         """
         connection.close()
 
@@ -1709,15 +1665,16 @@ class Transport(base.Transport):
         For each drained message, the message is called to the appropriate
         callback. Callbacks are organized by queue name.
 
-        :param connection: The :class:`Connection` that contains the
-            callbacks, indexed by queue name, which will be called by this
-            method.
-        :type connection: Connection
+        :param connection: The :class:`~kombu.transport.qpid.Connection` that
+            contains the callbacks, indexed by queue name, which will be called
+            by this method.
+        :type connection: kombu.transport.qpid.Connection
         :keyword timeout: The timeout that limits how long this method will
             run for. The timeout could interrupt a blocking read that is
             waiting for a new message, or cause this method to return before
             all messages are drained. Defaults to 0.
         :type timeout: int
+
         """
         start_time = time.time()
         elapsed_time = -1
@@ -1734,18 +1691,19 @@ class Transport(base.Transport):
         raise socket.timeout()
 
     def create_channel(self, connection):
-        """Create and return a :class:`Channel`.
+        """Create and return a :class:`~kombu.transport.qpid.Channel`.
 
-        Creates a new :class:`Channel`, and append the :class:`Channel` to the
-        list of channels known by the :class:`Connection`. Once the new
-        :class:`Channel` is created, it is returned.
+        Creates a new channel, and appends the channel to the
+        list of channels known by the Connection.  Once the new
+        channel is created, it is returned.
 
         :param connection: The connection that should support the new
-            :class:`Channel`.
-        :type connection: Connection
+            :class:`~kombu.transport.qpid.Channel`.
+        :type connection: kombu.transport.qpid.Connection
 
         :return: The new Channel that is made.
-        :rtype: :class:`Channel`.
+        :rtype: :class:`kombu.transport.qpid.Channel`.
+
         """
         channel = connection.Channel(connection, self)
         connection.channels.append(channel)
@@ -1760,6 +1718,7 @@ class Transport(base.Transport):
 
         :return: A dict containing the default parameters.
         :rtype: dict
+
         """
         return {
             'hostname': 'localhost',
@@ -1767,9 +1726,7 @@ class Transport(base.Transport):
         }
 
     def __del__(self):
-        """
-        Ensure file descriptors opened in __init__() are closed.
-        """
+        """Ensure file descriptors opened in __init__() are closed."""
         for fd in (self.r, self._w):
             try:
                 os.close(fd)